Bug fix
An OpenGL was being posted when the DVR panel was disabled. The GLWindow was
not setting the OpenGL context before deleting the renders. When the
ShaderProgram attempted to delete its underlying glShader and glProgram
objects, it attempted to do so in in the wrong graphics context.

This had not been an issue before the introduction of the opengl-based transfer
function. The Shader had luckly been in the correct context when deleting, but
the introduction of a new context shifted the luck...

A "makeCurrent()" call has been added to GLWindow::deleteRenderer(...), and
a printOpenGLError() have been added to the ShaderProgram destructor.