Lighting bug fix
The DVR lighting fragment shader was using the texture coordinate as the
position in the lighting calculation (a kludgy optimization). This caused
problems with bricked volumes, since each brick has a local texture coordinate
system. A vertex shader has been added to the lighting program to pass a
varying position vector into the fragment shader, providing the actual
fragment position for the lighting calculation.