Lighting bug fix (part 2)
The light triplet (x,y,z) specified in the viewpoint tab is taken by the
rendering classes as a light direction vector (relative to the camera). The
recent DVR lighting bug fix treated this triplet as a position (in the
camera's coordinate system). This has been corrected and made consistent with
the other renderers. The DVR now treats this triplet as a light direction.