Pre-integration rendering bug (un-filed)
The front and back of the proxy slab were being calculated in texture coordinate space,assuming 1:1 correspondence with volume extents. This was problematic in a variety of circumstances (e.g., anisotropic grids, bricked volumes, ...).

A Texture matrix is now initialized that allows the vertex shader to convert object coordinates into texture coordinates. The vertex shader calculates the back of the proxy slab in object space and then transformed into texture space for the texture lookup in the fragment shader.