Spherical Lighting
Adds lighting capabilities to the spherical volume render.

The gradient calculation is suspect. The light position and/or gradient
direction seems incorrect in some orientations. There is likely an error
somewhere in the multiple spherical <-> cartesian point conversions, and
the change of basis for the sperical to cartesian gradient vector. This
needs to be looked into, but this gives a fairly reasonable first pass.