Clean up
There was an ugly and redundant circular dependency between MapperFunction,
OpacityMap/Colormap and RenderParams. MapperFunction maintained a pointer to
the current RenderParams, so it could pass it to it's child color and opacity
maps, which would use the RenderParams to look up the current function bounds,
which would call the parent MapperFunction. Ugh.

The dependency on RenderParams has been removed from OpacityMap, and ColorMap.
Boundary information is now taken directly from the parent MapperFunction.