Added timestep-dependent SetExtents() and GetExtents() methods.
Made handling of required and optional Metadata elements consistent. In
particular, a Get method for a required element will generate an error
if the element is not defined. A Get method for an optional element (e.g. the
time step dependent GetExtents method) will return an empty vector, but
not throw an error if the element does exist.